requirements.txt
numpy==1.26.4
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.1.3
pytorch-msssim==1.0.0
pytorchvideo==0.1.5
grpcio==1.57.0
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
pillow==10.4.0
pillow_heif==0.15.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
transformers==4.43.3
tqdm==4.65.0
moviepy==1.0.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
timm==1.0.15
